refactor: 게임 전체를 컨트롤하는 로직을 분리
